+++ Osiris/Config.h
@@ -37,7 +37,7 @@ public:
         bool bunnyHop{ false };
         bool disablePostProcessing{ false };
         std::string skybox{ xorstr_("Default") };
-        float viewmodelFov{ 60.0f };
+        int viewmodelFov{ 60 };
     } misc;
 
 private:
